refactor: isolando criação e listagem de dados
