+++ Program.cs
@@ -1,7 +1,9 @@
 ﻿using LabManager.Database;
+using LabManager.Repositories;
 using Microsoft.Data.Sqlite;
 
 var databaseSetup = new DatabaseSetup(); 
+var computerRepository = new ComputerRepository();
 
 var modelName = args[0];
 var modelAction = args[1];
@@ -10,20 +12,10 @@ if (modelName == "Computer")
 {
     if (modelAction == "List")
     {
-        var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open();
-        var command = connection.CreateCommand();
-
-        command.CommandText = $"SELECT * FROM Computer;";
-
-        var reader = command.ExecuteReader();
-
-        while (reader.Read())
+        foreach(var computer in computerRepository.GetAll())
         {
-            Console.WriteLine($"{reader.GetInt32(0)} {reader.GetString(1)} {reader.GetString(2)}");
+            Console.WriteLine($"{computer.ID}, {computer.Ram}, {computer.Processor}");
         }
-
-        connection.Close();
     }
     else if (modelAction == "New")
     {
